pyproject.toml
[build-system]
requires = ["setuptools>=68", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "keyboard-familiar"
version = "0.1.0"
description = "Keyboard Familiar Harness modular runtime"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
  "pydantic>=2.7",
  "typer>=0.12",
  "pyyaml>=6.0",
]

[project.optional-dependencies]
dev = ["pytest>=8.0", "pytest-asyncio>=0.23"]

[project.scripts]
familiar = "familiar.cli:app"

[tool.setuptools.packages.find]
include = ["familiar*"]
exclude = ["config*"]
namespaces = false
